pyproject.toml
[tool.poetry]
name = "simple-prod-api"
version = "0.1.0"
description = ""
authors = ["Abinav R <abinav.rvkrish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pre-commit = "^3.3.3"
black = "^23.7.0"
bandit = "^1.7.5"
radon = "^6.0.1"
flake8 = "^6.1.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
